lambda/image_generator/requirements.txt
boto3>=1.34
openai>=1.0
Pillow>=10.0
requests>=2.31
